Refactor DB and remote data sources. Change Room's LiveData to Flowable. Add units and language injection into WeatherApi interceptor.
