+++ app/src/main/java/ru/exwhythat/yather/data/local/ForecastWeatherDao.java
@@ -1,0 +1,34 @@
+package ru.exwhythat.yather.data.local;
+
+import android.arch.persistence.room.Dao;
+import android.arch.persistence.room.Insert;
+import android.arch.persistence.room.OnConflictStrategy;
+import android.arch.persistence.room.Query;
+
+import java.util.List;
+
+import io.reactivex.Flowable;
+import ru.exwhythat.yather.data.local.entities.ForecastWeather;
+
+/**
+ * Created by exwhythat on 8/5/17.
+ */
+
+@Dao
+public interface ForecastWeatherDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    long insert(ForecastWeather forecastWeather);
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    long[] insertAll(List<ForecastWeather> forecastWeatherList);
+
+    @Query("SELECT * FROM forecastweather")
+    Flowable<List<ForecastWeather>> getAll();
+
+    @Query("SELECT * FROM forecastweather WHERE apiCityId = :cityId")
+    Flowable<List<ForecastWeather>> getForCity(int cityId);
+
+    @Query("DELETE FROM forecastweather WHERE apiCityId = :cityId")
+    int deleteForCity(int cityId);
+}
