+++ app/src/main/java/ru/exwhythat/yather/data/local/entities/City.java
@@ -1,0 +1,82 @@
+package ru.exwhythat.yather.data.local.entities;
+
+import android.arch.persistence.room.Entity;
+import android.arch.persistence.room.PrimaryKey;
+
+/**
+ * Created by exwhythat on 8/4/17.
+ */
+
+@Entity
+public class City {
+
+    @PrimaryKey private int apiCityId;
+    private double latitude;
+    private double longitude;
+    private String name;
+
+    public City(int apiCityId, double latitude, double longitude, String name) {
+        this.apiCityId = apiCityId;
+        this.latitude = latitude;
+        this.longitude = longitude;
+        this.name = name;
+    }
+
+    public int getApiCityId() {
+        return apiCityId;
+    }
+
+    public void setApiCityId(int apiCityId) {
+        this.apiCityId = apiCityId;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(double longitude) {
+        this.longitude = longitude;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof City)) return false;
+
+        City city = (City) o;
+
+        if (apiCityId != city.apiCityId) return false;
+        if (Double.compare(city.latitude, latitude) != 0) return false;
+        if (Double.compare(city.longitude, longitude) != 0) return false;
+        return name.equals(city.name);
+    }
+
+    @Override
+    public int hashCode() {
+        int result;
+        long temp;
+        result = apiCityId;
+        temp = Double.doubleToLongBits(latitude);
+        result = 31 * result + (int) (temp ^ (temp >>> 32));
+        temp = Double.doubleToLongBits(longitude);
+        result = 31 * result + (int) (temp ^ (temp >>> 32));
+        result = 31 * result + name.hashCode();
+        return result;
+    }
+}
