+++ app/src/main/java/ru/exwhythat/yather/data/remote/WeatherApiConstants.java
@@ -1,0 +1,10 @@
+package ru.exwhythat.yather.data.remote;
+
+/**
+ * Created by exwhythat on 8/6/17.
+ */
+
+public class WeatherApiConstants {
+    public static final String WEATHER_API_KEY = "4ac1ae796a583e819faf2f2bec3f0aed";
+    public static final String DEFAULT_UNITS = "metric";
+}
