+++ app/src/main/java/ru/exwhythat/yather/data/remote/model/BaseResponse.java
@@ -1,0 +1,9 @@
+package ru.exwhythat.yather.data.remote.model;
+
+/**
+ * Created by Sinjvf on 16.07.2017.
+ * Super class for all responses
+ */
+
+public class BaseResponse {
+}
