+++ app/src/main/java/ru/exwhythat/yather/data/remote/model/City.java
@@ -1,0 +1,52 @@
+package ru.exwhythat.yather.data.remote.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * Created by exwhythat on 8/3/17.
+ */
+
+public class City {
+
+    @SerializedName("id") @Expose
+    private int id;
+    @SerializedName("name") @Expose
+    private String name;
+    @SerializedName("coord") @Expose
+    private Coord coord;
+    @SerializedName("country") @Expose
+    private String countryCode;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Coord getCoord() {
+        return coord;
+    }
+
+    public void setCoord(Coord coord) {
+        this.coord = coord;
+    }
+
+    public String getCountryCode() {
+        return countryCode;
+    }
+
+    public void setCountryCode(String countryCode) {
+        this.countryCode = countryCode;
+    }
+}
