+++ app/src/main/java/ru/exwhythat/yather/data/remote/model/Weather.java
@@ -1,0 +1,53 @@
+package ru.exwhythat.yather.data.remote.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * Created by Sinjvf on 16.07.2017.
+ * Data class have simple weather description
+ */
+
+public class Weather {
+
+    @SerializedName("id") @Expose
+    private int id;
+    @SerializedName("main") @Expose
+    private String main;
+    @SerializedName("description") @Expose
+    private String description;
+    @SerializedName("icon") @Expose
+    private String icon;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getMain() {
+        return main;
+    }
+
+    public void setMain(String main) {
+        this.main = main;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getIcon() {
+        return icon;
+    }
+
+    public void setIcon(String icon) {
+        this.icon = icon;
+    }
+}
