--- app/src/main/java/ru/exwhythat/yather/network/weather/parts/BaseResponse.java
@@ -1,9 +1,0 @@
-package ru.exwhythat.yather.network.weather.parts;
-
-/**
- * Created by Sinjvf on 16.07.2017.
- * Super class for all responses
- */
-
-public class BaseResponse {
-}
