--- app/src/main/java/ru/exwhythat/yather/network/weather/parts/Clouds.java
@@ -1,23 +1,0 @@
-package ru.exwhythat.yather.network.weather.parts;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-/**
- * Created by Sinjvf on 15.07.2017.
- * Data class describes clouds
- */
-
-public class Clouds {
-
-    @SerializedName("all") @Expose
-    private int all;
-
-    public int getAll() {
-        return all;
-    }
-
-    public void setAll(int all) {
-        this.all = all;
-    }
-}
