--- app/src/main/java/ru/exwhythat/yather/network/weather/parts/Coord.java
@@ -1,33 +1,0 @@
-package ru.exwhythat.yather.network.weather.parts;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-/**
- * Created by Sinjvf on 16.07.2017.
- * Data class describes coordinates
- */
-
-public class Coord {
-
-    @SerializedName("lon") @Expose
-    private double lon;
-    @SerializedName("lat") @Expose
-    private double lat;
-
-    public double getLon() {
-        return lon;
-    }
-
-    public void setLon(double lon) {
-        this.lon = lon;
-    }
-
-    public double getLat() {
-        return lat;
-    }
-
-    public void setLat(double lat) {
-        this.lat = lat;
-    }
-}
