--- app/src/main/java/ru/exwhythat/yather/network/weather/parts/ForecastList.java
@@ -1,104 +1,0 @@
-package ru.exwhythat.yather.network.weather.parts;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-/**
- * Created by exwhythat on 8/3/17.
- */
-
-public class ForecastList {
-
-    @SerializedName("dt") @Expose
-    private int dt;
-    @SerializedName("main") @Expose
-    private Main main;
-    @SerializedName("weather") @Expose
-    private List<Weather> weather = null;
-    @SerializedName("clouds") @Expose
-    private Clouds clouds;
-    @SerializedName("wind") @Expose
-    private Wind wind;
-    @SerializedName("rain") @Expose
-    private Rain rain;
-    @SerializedName("snow") @Expose
-    private Snow snow;
-    @SerializedName("sys") @Expose
-    private Sys sys;
-    @SerializedName("dt_txt") @Expose
-    private String dtTxt;
-
-    public int getDt() {
-        return dt;
-    }
-
-    public void setDt(int dt) {
-        this.dt = dt;
-    }
-
-    public Main getMain() {
-        return main;
-    }
-
-    public void setMain(Main main) {
-        this.main = main;
-    }
-
-    public List<Weather> getWeather() {
-        return weather;
-    }
-
-    public void setWeather(List<Weather> weather) {
-        this.weather = weather;
-    }
-
-    public Clouds getClouds() {
-        return clouds;
-    }
-
-    public void setClouds(Clouds clouds) {
-        this.clouds = clouds;
-    }
-
-    public Wind getWind() {
-        return wind;
-    }
-
-    public void setWind(Wind wind) {
-        this.wind = wind;
-    }
-
-    public Rain getRain() {
-        return rain;
-    }
-
-    public void setRain(Rain rain) {
-        this.rain = rain;
-    }
-
-    public Snow getSnow() {
-        return snow;
-    }
-
-    public void setSnow(Snow snow) {
-        this.snow = snow;
-    }
-
-    public Sys getSys() {
-        return sys;
-    }
-
-    public void setSys(Sys sys) {
-        this.sys = sys;
-    }
-
-    public String getDtTxt() {
-        return dtTxt;
-    }
-
-    public void setDtTxt(String dtTxt) {
-        this.dtTxt = dtTxt;
-    }
-}
