--- app/src/main/java/ru/exwhythat/yather/network/weather/parts/Snow.java
@@ -1,18 +1,0 @@
-package ru.exwhythat.yather.network.weather.parts;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-/**
- * Created by exwhythat on 8/3/17.
- */
-
-public class Snow {
-
-    @SerializedName("3h") @Expose
-    public double lastThreeHours;
-
-    public double getLastThreeHours() {
-        return lastThreeHours;
-    }
-}
